# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.ApproveIncrementRequest:
    properties:
      effective_date:
        type: string
      increment_amount:
        type: number
    type: object
...
  handlers.CreateFloorRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  handlers.CreateIncrementRequest:
    properties:
      company_id:
        type: string
      effective_date:
        type: string
      employee_id:
        type: string
      increment_amount:
        minimum: 1
        type: number
      remarks:
        type: string
    required:
    - company_id
    - effective_date
    - employee_id
    - increment_amount
    type: object
...
  handlers.UpdateFloorRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  utils.PaginatedResponse:
    properties:
      data: {}
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:5000
info:
  contact:
    email: support@peoplehub.com
    name: PeopleHub Team
  description: Enterprise HR Management & Payroll System API.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: PeopleHub API
  version: "1.1"
paths:
...
  /floors:
    get:
      description: Get all floors
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List floors
      tags:
      - Floors
    post:
      consumes:
      - application/json
      description: Create a new floor
      parameters:
      - description: Floor details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateFloorRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create floor
      tags:
      - Floors
  /floors/{id}:
    delete:
      description: Soft delete a floor
      parameters:
      - description: Floor ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete floor
      tags:
      - Floors
    get:
      description: Get a floor by its ID
      parameters:
      - description: Floor ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get floor by ID
      tags:
      - Floors
    put:
      consumes:
      - application/json
      description: Update an existing floor
      parameters:
      - description: Floor ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated floor details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateFloorRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update floor
      tags:
      - Floors
...
  /salary/increments:
    get:
      description: Get all salary increments for a company
      parameters:
      - description: Company ID
        in: query
        name: company_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List salary increments
      tags:
      - Salary
    post:
      consumes:
      - application/json
      description: Apply a salary increment for an employee
      parameters:
      - description: Increment details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateIncrementRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create salary increment
      tags:
      - Salary
  /salary/increments/{id}/approve:
    put:
      consumes:
      - application/json
      description: Approve a pending increment and update employee salary
      parameters:
      - description: Increment ID
        in: path
        name: id
        required: true
        type: string
      - description: Optional override values
        in: body
        name: request
        schema:
          $ref: '#/definitions/handlers.ApproveIncrementRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve salary increment
      tags:
      - Salary
  /salary/increments/{id}/reject:
    put:
      consumes:
      - application/json
      description: Reject a pending increment request
      parameters:
      - description: Increment ID
        in: path
        name: id
        required: true
        type: string
      - description: Rejection reason
        in: body
        name: request
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Reject salary increment
      tags:
      - Salary
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"